+++ src/gui/elements/NameBox.ts
@@ -8,20 +8,20 @@ export default class NameBox extends Sprite {
     text: Label
     game: RJS
     config: {
-        id: string,
-        asset: string,
-        x: number,
-        y: number,
-        transition?: string,
-        tintStyle?: string,
+        id: string;
+        asset: string;
+        x: number;
+        y: number;
+        transition?: string;
+        tintStyle?: string;
         text: {
-            x: number,
-            y: number,
-            width: number,
-            height: number,
-            style: any,
-            lineSpacing: number
-        }
+            x: number;
+            y: number;
+            width: number;
+            height: number;
+            style: any;
+            lineSpacing: number;
+        };
     }
 
     constructor(game: RJS, config) {
@@ -46,14 +46,14 @@ export default class NameBox extends Sprite {
             this.text.fill = color;
         }
         this.visible = true;
-        let transition = this.game.screenEffects.transition.get(this.config.transition);
+        const transition = this.game.screenEffects.transition.get(this.config.transition);
         await transition(null,this);
     }
 
     async hide(transitionName?){
         if (!this.visible) return;
         if (!transitionName) transitionName = this.config.transition;
-        let transition = this.game.screenEffects.transition.get(transitionName);
+        const transition = this.game.screenEffects.transition.get(transitionName);
         await transition(this,null);
         this.visible = false;
         this.text.setText('', true);