+++ src/managers/StoryManager.ts
@@ -138,12 +138,12 @@ export default class StoryManager implements StoryManagerInterface<Group> {
     // removes everything from screen immediately
     clearScene(): void{
         this.game.control.execStack.clear();
-        this.game.gui.hud.clear("CUT");
+        this.game.gui.hud.clear('CUT');
         this.game.control.waitForClick = false;
-        this.game.managers.character.hideAll("CUT");
+        this.game.managers.character.hideAll('CUT');
         this.game.managers.audio.stopAll()
-        this.game.managers.cgs.hideAll("CUT");
-        this.game.managers.background.hide(null,"CUT");
+        this.game.managers.cgs.hideAll('CUT');
+        this.game.managers.background.hide(null,'CUT');
         this.game.screenEffects.ambient.CLEAR();
         this.currentScene = [];
     }
@@ -151,7 +151,7 @@ export default class StoryManager implements StoryManagerInterface<Group> {
     async startScene(name: string) {
         await this.game.gui.hud.clear();
         if (this.game.setup.lazyloading){
-            // load scene or episode assets (not loaded yet) 
+            // load scene or episode assets (not loaded yet)
             await this.assetLoader.loadScene(name);
         }
         this.game.control.execStack.replace(name);
@@ -182,8 +182,11 @@ export default class StoryManager implements StoryManagerInterface<Group> {
             this.actorsIndex[actor] = 'sfx';
             return 'sfx';
         }
-        this.actorsIndex[actor] = 'cgs';
-        return 'cgs';
+        if (this.game.managers.cgs.isCGS(actor)){
+            this.actorsIndex[actor] = 'cgs';
+            return 'cgs';
+        }
+        return null;
     }
 
     getManagerByActorType (type: string): RJSSpriteManagerInterface {
@@ -197,10 +200,10 @@ export default class StoryManager implements StoryManagerInterface<Group> {
 
     parseAction(actionRaw): StoryAction {
         const keyParams = Object.keys(actionRaw)[0].split(' ');
-        let actionType = (keyParams[1] === 'says') ? 'say' : keyParams[0];
+        const actionType = (keyParams[1] === 'says') ? 'say' : keyParams[0];
         const ActionClass = this.actionFactory[actionType];
         if (ActionClass){
-            return new ActionClass(this.game,actionType,actionRaw);    
+            return new ActionClass(this.game,actionType,actionRaw);
         }
         return null;
     }
@@ -214,7 +217,7 @@ export default class StoryManager implements StoryManagerInterface<Group> {
             return this.game.resolveAction();
         }
         if (this.game.config.debugMode){
-            console.log("Executing action: "+storyAction.actionType);
+            console.log('Executing action: '+storyAction.actionType);
             console.log(Object.getOwnPropertyNames(storyAction));
         }
         await this.game.checkPlugins('onAction',[storyAction]);