+++ src/states/PreloadStory.ts
@@ -5,8 +5,8 @@ import RJSState from './RJSState';
 import RJSLoadingScreen from '../gui/elements/RJSLoadingScreen';
 
 class PreloadStory extends RJSState {
-    loadingScreen:RJSLoadingScreen
-    readyToStart: boolean = false;
+    loadingScreen: RJSLoadingScreen
+    readyToStart = false;
 
     constructor() {
         super();
@@ -78,14 +78,14 @@ class PreloadStory extends RJSState {
         // preload background music
         if ('music' in this.game.setup) {
             for (const music in this.game.setup.music) {
-               preloadAudio(music,"music",this.game);
+               preloadAudio(music,'music',this.game);
             }
         }
 
         // preload sfx
         if ('sfx' in this.game.setup) {
             for (const sfx in this.game.setup.sfx) {
-               preloadAudio(sfx,"sfx",this.game);
+               preloadAudio(sfx,'sfx',this.game);
             }
         }
 
@@ -116,7 +116,7 @@ class PreloadStory extends RJSState {
         } else {
             this.readyToStart = true;
         }
-        
+
     }
 
     initGame(){